# Performance improvement for `cupy.var` complex inputs