Rel #2411.
Speed up `cupy.var` for complex inputs by defining custom kernel for complex value.