@niboshi 2019-09-20
Can you provide benchmark results?