@grlee77 2019-09-23
> Can you provide benchmark results?

This PR seems to be almost a factor of two improvement over current master.

Using `cupy.std` on a complex64 arrays:

with 100,000 elements:
```
    master: 314 µs ± 12.4 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
    this PR: 168 µs ± 12.4 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
```

with 100,000,000 elements:
```
    master: 205 ms ± 55.4 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
    this PR: 113 ms ± 8.47 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

Similar improvement occurs for `cupy.var`