@asi1024 2019-09-24
I removed redundant memory copies of the output array and measured performances with the following 2 types of inputs.
- var_float: `shape=(1000, 1000), dtype='float32'`
- var_complex: `shape=(1000, 1000), dtype='complex64'`

- `master` (`3d5d5d3c4`)
```
var_float      :    49.627 us   +/-53.548 (min:   46.125) us    806.001 us   +/-48.474 (min:  783.776) us
var_complex    :   118.108 us   +/-18.705 (min:  112.390) us   3062.695 us   +/-14.864 (min: 3044.384) us
```

- This PR (`db13a973e`)
```
var_float      :    58.410 us   +/- 8.308 (min:   56.412) us    807.499 us   +/-25.925 (min:  791.104) us
var_complex    :    59.996 us   +/- 9.603 (min:   57.613) us   1677.435 us   +/- 9.098 (min: 1653.408) us
```