@asi1024 2019-09-24
Add a commit for more performance improvement.
`HEAD` (`eef646801`)
```
var_float      :    54.315 us   +/- 7.526 (min:   52.076) us    791.380 us   +/- 6.337 (min:  772.160) us
var_complex    :    56.022 us   +/- 8.459 (min:   53.555) us   1671.560 us   +/- 7.467 (min: 1657.728) us
```

Re-measured in `master` because the previous `var_float` result have too large divergence (`+/-53.548`)
```
var_float      :    54.669 us   +/- 8.657 (min:   49.902) us    792.045 us   +/- 6.150 (min:  772.128) us
var_complex    :   130.345 us   +/-15.724 (min:  122.815) us   3064.546 us   +/-10.047 (min: 3040.160) us
```