@niboshi 2019-09-25
@asi1024 @grlee77
Thank you for the benchmark results.

@asi1024 Please fix the test failure.